examples/new_postgres_api/requirements.txt
# Core
nzrapi>=2.0.0
python-dotenv>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Database
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.4
asyncpg>=0.24.0

# Web
uvicorn[standard]>=0.15.0
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5

# Development
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.19.0
black>=21.9b0
flake8>=4.0.1
mypy>=0.910
pytest-cov>=2.12.1

# Documentation
mkdocs>=1.2.3
mkdocs-material>=7.3.6

# Optional (for production)
gunicorn>=20.1.0
uvloop>=0.16.0
httptools>=0.2.1
